roles/home_automation/files/wol-processor/requirements.txt
paho-mqtt
wakeonlan
